pyproject.toml
[project]
name = "alfred"
version = "0.1.0"
description = ""
authors = [
    {name = "Joel Guerrero",email = "joelgur10@hotmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django (>=5.2,<6.0)",
    "djangorestframework (>=3.16.0,<4.0.0)",
    "djangorestframework-simplejwt (>=5.5.0,<6.0.0)",
    "faker (>=37.1.0,<38.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "drf-spectacular (>=0.28.0,<0.29.0)",
    "drf-spectacular-sidecar (>=2025.4.1,<2026.0.0)"
]

[tool.poetry]
package-mode = false

[tool.poetry.group.dev.dependencies]
black = "^25.1.0"
isort = "^6.0.1"
flake8 = "^7.2.0"
pre-commit = "^4.2.0"
coverage = "^7.8.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120
target-version = ['py310']

[tool.isort]
profile = "black"

[tool.flake8]
max-line-length = 120
extend-ignore = ["E203", "W503"]